+++ Library/DaylightModelController.swift
@@ -1,14 +1,36 @@
+import CoreLocation
 import UIKit
 
-class DaylightModelController {
-    private var location: Location
+protocol DaylightModelControllerObserver: AnyObject {
+    func daylightModelControllerDidUpdate(_ controller: DaylightModelController)
+}
 
-    init(location: Location) {
-        self.location = location
+class DaylightModelController {
+    private var location: Location? {
+        didSet {
+            updateObservers()
+        }
+    }
+
+    lazy var locationTracker: LocationTracker = {
+        let tracker = LocationTracker()
+        tracker.delegate = self
+
+        Timer.scheduledTimer(timeInterval: 30, target: self, selector: #selector(self.updateObservers), userInfo: nil, repeats: true)
+
+        return tracker
+    }()
+
+    init() {
+        self.locationTracker.locateIfPossible()
+    }
+
+    @objc func updateObservers() {
+        
     }
 }
 
-extension  DaylightModelController {
+extension DaylightModelController {
     var sunPhase: SunTime.SunPhase {
         return location.sunTime.sunPhase
     }
@@ -28,5 +50,29 @@ extension  DaylightModelController {
     var attributedMessage: NSAttributedString {
         let message = Message(for: Date(), coordinates: location.coordinates)
         return message.attributedString(textColor: secondaryColor.withAlphaComponent(0.6), highlightColor: secondaryColor)
+    }
+}
+extension DaylightModelController {
+    func addObserver(_ observer: DaylightModelControllerObserver) {
+
+    }
+}
+
+extension DaylightModelController: LocationTrackerDelegate {
+
+    func locationTracker(_ locationTracker: LocationTracker, didFailWith error: Error) {
+        guard Location.current == nil else { return }
+
+        let isUnexpectedError = (error as NSError).code != 0
+        if isUnexpectedError {
+            let alertController = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
+            alertController.addAction(UIAlertAction(title: "Dismiss", style: .default, handler: nil))
+            self.present(alertController, animated: true, completion: nil)
+        }
+    }
+
+    func locationTracker(_ locationTracker: LocationTracker, didFindLocation placemark: CLPlacemark) {
+        self.location = Location(placemark: placemark)
+
     }
 }
