+++ iOS/Source/MainController.swift
@@ -1,21 +1,15 @@
-import CoreLocation
 import SweetUIKit
 import UIKit
 
 class MainController: UIViewController {
+    private let dayLightModelController: DaylightModelController()
+
     lazy var informationController: InformationController = {
         let informationController = InformationController()
         informationController.modalTransitionStyle = .crossDissolve
         informationController.delegate = self
 
         return informationController
-    }()
-
-    lazy var locationTracker: LocationTracker = {
-        let tracker = LocationTracker()
-        tracker.delegate = self
-
-        return tracker
     }()
 
     fileprivate var rootView: MainView {
@@ -39,17 +33,13 @@ class MainController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        self.locationTracker.locateIfPossible()
-        self.updateInterface()
-
-        Timer.scheduledTimer(timeInterval: 30, target: self, selector: #selector(self.updateInterface), userInfo: nil, repeats: true)
-
-        Notifier.cancelAllNotifications()
-        if Settings.areNotificationsEnabled {
-            if let location = Location.current {
-                Notifier.scheduleNotifications(for: location)
-            }
-        }
+        //This should be done in the model
+//        Notifier.cancelAllNotifications()
+//        if Settings.areNotificationsEnabled {
+//            if let location = Location.current {
+//                Notifier.scheduleNotifications(for: location)
+//            }
+//        }
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -59,12 +49,6 @@ class MainController: UIViewController {
         UIView.animate(withDuration: TransitionDuration, animations: {
             self.rootView.alpha = 1.0
         })
-    }
-
-    @objc func updateInterface() {
-        if let location = Location.current {
-            self.rootView.updateInterface(controller: DaylightModelController(location: location))
-        }
     }
 }
 
@@ -86,26 +70,6 @@ extension MainController: MainViewDelegate {
         let activityController = UIActivityViewController(activityItems: [screenshot, "Made with #daylightapp."], applicationActivities: nil)
         activityController.excludedActivityTypes = [UIActivityType.airDrop]
         self.present(activityController, animated: true, completion: nil)
-    }
-}
-
-extension MainController: LocationTrackerDelegate {
-
-    func locationTracker(_ locationTracker: LocationTracker, didFailWith error: Error) {
-        guard Location.current == nil else { return }
-
-        let isUnexpectedError = (error as NSError).code != 0
-        if isUnexpectedError {
-            let alertController = UIAlertController(title: "Error", message: error.localizedDescription, preferredStyle: .alert)
-            alertController.addAction(UIAlertAction(title: "Dismiss", style: .default, handler: nil))
-            self.present(alertController, animated: true, completion: nil)
-        }
-    }
-
-    func locationTracker(_ locationTracker: LocationTracker, didFindLocation placemark: CLPlacemark) {
-        Location.current = Location(placemark: placemark)
-
-        self.updateInterface()
     }
 }
